added ghost and timeout
The --ghost option creates "ghost" versions (i.e. empty directories) of all mount points defined in the configuration file regardless of whether or not any of the file systems are currently mounted there. Without it, unmounted filesystems will not appear at all and you will have to look in /etc/auto.network if you forget the name of a mount point.